feat: new library
Implement a new, updated library for better documentation and to
allow for more robust error handling and notification methods.
