+++ package.json
@@ -38,10 +38,9 @@
     "prettier": "^2.2.1"
   },
   "dependencies": {
-    "@types/twitter": "^1.7.0",
     "dotenv": "^8.2.0",
     "node-fetch": "^2.6.1",
-    "twitter": "^1.7.1"
+    "twitter-lite": "^1.1.0"
   },
   "lint-staged": {
     "*.ts": [
