+++ src/handlers/onEnd.ts
@@ -1,0 +1,34 @@
+import { WebhookBodyInt } from "../interfaces/WebhookBodyInt";
+import { WebhookEmbedInt } from "../interfaces/WebhookEmbedInt";
+import { getWebhook } from "../modules/webhook";
+import fetch from "node-fetch";
+
+export const onEnd = async (): Promise<void> => {
+  const webhook = await getWebhook();
+
+  const originalEmbed: WebhookEmbedInt = {
+    title: "Twitter Monitor is shutting down!",
+    description: "Something is broken! <@!465650873650118659>, check the logs!",
+  };
+
+  const webhookBody: WebhookBodyInt = {
+    content: "A wild system message appeared!",
+    username: "Twitter Bot",
+    avatar_url: "https://cdn.nhcarrigan.com/content/profile.jpg",
+    allowed_mentions: {
+      parse: [],
+    },
+    embeds: [originalEmbed],
+  };
+
+  const sentData = await fetch(webhook, {
+    method: "post",
+    body: JSON.stringify(webhookBody),
+    headers: { "Content-Type": "application/json" },
+  });
+
+  if (!sentData || sentData.status !== 204) {
+    console.error("Failed to send tweet:");
+    console.info("Twitter Monitor is online!");
+  }
+};
