+++ src/index.ts
@@ -1,17 +1,13 @@
-import fetch from "node-fetch";
-import { TweetInt } from "./interfaces/TweetInt";
-import { WebhookBodyInt } from "./interfaces/WebhookBodyInt";
-import { WebhookEmbedInt } from "./interfaces/WebhookEmbedInt";
+import { onData } from "./handlers/onData";
+import { onEnd } from "./handlers/onEnd";
+import { onStart } from "./handlers/onStart";
 import { login } from "./modules/login";
-import { getWebhook } from "./modules/webhook";
 
 (async () => {
   const Client = login();
 
-  const WebhookUrl = await getWebhook();
-
   try {
-    const data = await Client.get("account/verify_credentials", {});
+    const data = await Client.get("account/verify_credentials");
     console.log(
       `Client has logged in! User: ${data.screen_name} - ID: ${data.id_str}`
     );
@@ -21,91 +17,13 @@ import { getWebhook } from "./modules/webhook";
     process.exit(1);
   }
 
-  Client.stream(
-    "statuses/filter",
-    { follow: "1261759785344499712" },
-    (stream) => {
-      stream.on("data", async (tweet: TweetInt) => {
-        console.log(tweet);
-        const {
-          in_reply_to_status_id_str: replyId,
-          text,
-          user,
-          created_at,
-          quoted_status: quote,
-        } = tweet;
-        const { profile_image_url: avatar } = user;
+  const stream = Client.stream("statuses/filter", {
+    follow: "1261759785344499712",
+  });
 
-        const originalEmbed: WebhookEmbedInt = {
-          title: "nhcarrigan tweeted!",
-          description: text,
-          url: `https://twitter.com/i/web/status/${tweet.id_str}`,
-          author: {
-            name: user.screen_name,
-            url: user.url,
-            icon_url: user.profile_image_url_https,
-          },
-          timestamp: new Date(created_at).toISOString(),
-        };
-
-        const webhookBody: WebhookBodyInt = {
-          content: "A wild tweet appeared!",
-          username: "Twitter Bot",
-          avatar_url: avatar,
-          allowed_mentions: {
-            parse: [],
-          },
-          embeds: [originalEmbed],
-        };
-
-        if (replyId) {
-          const replyData = await Client.get("statuses/show", {
-            id: replyId,
-          });
-          const typedData = replyData as TweetInt;
-
-          const replyEmbed: WebhookEmbedInt = {
-            title: "In reply to this tweet:",
-            description: typedData.text,
-            url: `https://twitter.com/i/web/status/${replyId}`,
-            author: {
-              name: typedData.user.screen_name,
-              url: typedData.user.url,
-              icon_url: typedData.user.profile_image_url_https,
-            },
-            timestamp: new Date(typedData.created_at).toISOString(),
-          };
-
-          webhookBody.embeds?.push(replyEmbed);
-        }
-
-        if (quote) {
-          const quoteEmbed: WebhookEmbedInt = {
-            title: "And quoted this tweet:",
-            description: quote.text,
-            url: `https://twitter.com/i/web/status/${quote.id_str}`,
-            author: {
-              name: quote.user.screen_name,
-              url: quote.user.url,
-              icon_url: quote.user.profile_image_url_https,
-            },
-            timestamp: new Date(quote.created_at).toISOString(),
-          };
-
-          webhookBody.embeds?.push(quoteEmbed);
-        }
-
-        const sentData = await fetch(WebhookUrl, {
-          method: "post",
-          body: JSON.stringify(webhookBody),
-          headers: { "Content-Type": "application/json" },
-        });
-
-        if (!sentData || sentData.status !== 204) {
-          console.error("Failed to send tweet:");
-          console.info(text);
-        }
-      });
-    }
-  );
+  stream.on("start", onStart);
+  stream.on("data", onData);
+  stream.on("ping", () => console.log("ping"));
+  stream.on("error", (error) => console.error(error));
+  stream.on("end", onEnd);
 })();
